pyproject.toml
[project]
name = "algoexpert"
version = "0.1.0"
description = "Algorithmic trading library"
authors = [{ name = "Oleg Alexandrov", email = "alexandrovoleg.ru@gmail.com" }]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=1.1.1",
    "websockets>=15.0.1",
    "pybit>=5.5.0",
]

[tool.pytest.ini_options]
pythonpath = [
  "src"
]

[tool.setuptools.packages.find]
where = [
  "src"
]

[project.optional-dependencies]
dev = [
    "mypy",
    "pytest",
    "ruff",
]

[tool.uv.workspace]
members = [
    ".",
    "examples/ema_expert_binance",
    "examples/ema_expert_bybit",
    "examples/websocket_test"
]